refactor(frontend): make the missing providers alert more buttony
